trebo/requirement.txt
Django==1.9
djangorestframework==3.4.7
mysqlclient==1.3.7
